+++ src/SharedAssemblyInfo.cs
@@ -3,12 +3,12 @@ using System.Reflection;
 
 [assembly: AssemblyCompanyAttribute("RazorEngine")]
 [assembly: AssemblyProductAttribute("RazorEngine")]
-[assembly: AssemblyCopyrightAttribute("RazorEngine Copyright © RazorEngine Project 2011-2015")]
-[assembly: AssemblyVersionAttribute("3.8.2")]
-[assembly: AssemblyFileVersionAttribute("3.8.2")]
-[assembly: AssemblyInformationalVersionAttribute("3.8.2")]
+[assembly: AssemblyCopyrightAttribute("RazorEngine Copyright © RazorEngine Project 2011-2016")]
+[assembly: AssemblyVersionAttribute("3.9.2")]
+[assembly: AssemblyFileVersionAttribute("3.9.2")]
+[assembly: AssemblyInformationalVersionAttribute("3.9.2")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "3.8.2";
+        internal const string Version = "3.9.2";
     }
 }