+++ doc/ReleaseNotes.md
@@ -1,4 +1,19 @@
-﻿### 3.8.2 / 4.3.2-beta1
+﻿### 3.9.2 / 4.4.2-rc1
+
+ * (PACKAGING) Fix version constraint, thanks campersau (https://github.com/Antaris/RazorEngine/pull/429)
+
+### 3.9.1 / 4.4.1-rc1
+
+ * (BUGFIX) dispose ReaderWriterLockSlim lock in ReferencesListForDynamicAssemblyResolution, thanks campersau (https://github.com/Antaris/RazorEngine/pull/411)
+ * (DOCS) update copyright year, thanks campersau (https://github.com/Antaris/RazorEngine/pull/412)
+ * (DOCS) Update template.cshtml, thanks (https://github.com/Antaris/RazorEngine/pull/413)
+ 
+### 3.9.0 / 4.4.0-rc1
+
+ * Fix memory leak in CompilerServiceBase by removing event handler in dispose, thanks campersau (https://github.com/Antaris/RazorEngine/pull/388)
+ * support AspNetCore.Razor 1.0.0, thanks campersau (https://github.com/Antaris/RazorEngine/pull/402)
+ 
+### 3.8.2 / 4.3.2-beta1
 
  * support models which implement ISerializable with IsolatedRazorEngineService (https://github.com/Antaris/RazorEngine/pull/376)
  * fix vb.net generic type parameter list (https://github.com/Antaris/RazorEngine/pull/375)